% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{devices}
\alias{devices}
\title{process sensor data from fit files}
\usage{
devices(
  device_info,
  cadenceFromPower = "auto",
  noncadPowermeters = c("ibike"),
  loud = FALSE
)
}
\arguments{
\item{device_info}{a dataframe of device info messages}

\item{cadenceFromPower}{cadence signal came from powermeter - specify
one of "always","never", or "auto"}

\item{noncadPowermeters}{a vector of manufacturer id strings for power meters which
do NOT provide the cadence signal, possibly overriding any cadence sensors}

\item{loud}{display summary of re/segmenting actions}
}
\value{
a list of hardware info
}
\description{
\code{devices} process sensor data from fit files.
}
\seealso{
\code{\link{read_ride}}
}
