% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{excludeCalibrate}
\alias{excludeCalibrate}
\title{exclude powerpod calibration sequence}
\usage{
excludeCalibrate(
  deltatime,
  watts,
  afterlast = TRUE,
  powerCalibrateAuto = TRUE,
  powerCalibrateTime = NA,
  maxCalDelta = 3,
  maxCalWatts = 100,
  maxminCalWatts = 3,
  minmaxCalWatts = 98,
  maxCalBurstSecs = 5,
  minBurstSize = 2000,
  loud = FALSE,
  ...
)
}
\arguments{
\item{deltatime}{timestamp vector}

\item{watts}{power vector}

\item{afterlast}{zap all power readings before last calibration}

\item{powerCalibrateAuto}{automatically find calibration signals if}

\item{powerCalibrateTime}{number of seconds to ignore in avgpower PostCal}

\item{maxCalDelta}{maximum stepsize in increasing sequence}

\item{maxCalWatts}{power value at end of calibration}

\item{maxminCalWatts}{largest start value for increasing sequence}

\item{minmaxCalWatts}{smallest end value for increasing sequence}

\item{maxCalBurstSecs}{window to look for huge values after sequence}

\item{minBurstSize}{maximum realistic power reading for burst elimination}

\item{loud}{print information about corrections}

\item{...}{arguments to other functions}
}
\value{
a vector containing cleaned power
}
\description{
\code{excludeCalibrate}  processes a gps track file to summarize the power data
}
\seealso{
\code{\link{read_ride}},\code{\link{excludeCalibrate}},
   \code{\link{repairSensorDropOut}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{repairPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}},
   \code{\link{statsTemp}}
}
