% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikeCadHr_options.R
\name{options to extract time and date from gps filenames}
\alias{options to extract time and date from gps filenames}
\title{filename-to-date parsing options for bikeCadHr functions}
\description{
filename-to-date parsing options for bikeCadHr functions
}
\details{
option names and defaults:
  Assumes names of .fit files from gps YYYY-MM-DD-HH-MM-SS.fit
  and names of .gpx files as exported from Garmin products
  M_D_YYYY H_MM_SS AM_history.gpx (where M,D, and H may be
  one or two digits)

  Modify these using \code{\link[base]{options}}
\itemize{
 \item{bCadHr.fit.fn.time.parse = "\%Y-\%m-\%d-\%H-\%M-\%S"}
 \item{bCadHr.fit.fn.lead = ""}
 \item{bCadHr.fit.fn.trail = ""}
 \item{bCadHr.gpx.fn.time.parse = "\%m_\%d_\%Y \%I_\%M_\%S \%p"}
 \item{bCadHr.gpx.fn.lead = ""}
 \item{bCadHr.gpx.fn.trail= "_history}
 }
}
