% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracktimes.R
\name{pedalTime}
\alias{pedalTime}
\title{return pedaling time for a track}
\usage{
pedalTime(
  trackdf,
  minCadence = 0,
  cadRequireRolling = TRUE,
  stopSpeed = 0,
  cadMinMove = 3,
  cadStartMaxDelay = 3,
  include,
  ...
)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{minCadence}{a numeric value below which the cadence is ignored}

\item{cadRequireRolling}{if TRUE, cadence value ignored if bike not moving}

\item{stopSpeed}{a numeric value below which the bike is considered stopped}

\item{cadMinMove}{a numeric value specifying the minimum duration of motion
to be included in pedaled time}

\item{cadStartMaxDelay}{a numeric value specifying the number of seconds
after beginning motion that pedaling is assumed for time (but not pedal
stroke accumulation).  This accounts for sensor startup lag.}

\item{include}{a vector specifying which points on the track to include}

\item{...}{parameters for other functions}
}
\value{
a vector of length one containing the elapsed time on the track for which
   include is TRUE
}
\description{
\code{pedalTime}  return the total time pedaling in a track
}
\seealso{
\code{\link{read_ride}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{totalTime}},
   \code{\link{rollingTime}}
}
