% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{processSegments}
\alias{processSegments}
\title{clean up and add start/stop segments to a track tibble, add flag for stopped}
\usage{
processSegments(trackdf, segInitIdleAggSecs = 60,
  segInitIdleAggMeters = 20, segBreakTimeMin = 3,
  nonsegTimeGapMax = 30, segSplitTimeStop = 3, segMinObs = 4,
  segMinMeters = 20, segMinSecs = 20, ignoreSegInfo = FALSE,
  loud = FALSE, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{segInitIdleAggSecs}{stops/restarts before this time has
elapsed are discarded and the ride starts at the last start prior to this}

\item{segInitIdleAggMeters}{stops/restarts within this distance
of the initial start are discarded as above}

\item{segBreakTimeMin}{(seconds) segments separated by less time
than this are joined}

\item{nonsegTimeGapMax}{(seconds) segemnts are split between datapoints
separated by more time than this}

\item{segSplitTimeStop}{segments are split at points where}

\item{segMinObs}{segments smaller than this are joined to closest
neighboring segment}

\item{segMinMeters}{segments shorter than this are joined with predecessor}

\item{segMinSecs}{(seconds) segments shorter than this are joined}

\item{ignoreSegInfo}{wipe existing segment information and use
\code{nonsegTimeGapMax} and \code{segSplitTimeStop} to
reconstruct segments before merging short ones}

\item{loud}{display summary of actions}

\item{...}{parameters for
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a tibble containing track data with start/stop segment data
}
\description{
\code{processSegments}  processes a gps track file to correct or add
   start/stop segment information like that generated by gps autopause.
   The sequence of processing is: eliminate stops that are too short,
   split segments where time between observations is too large, split
   segments where stopped for longer than threshold, then merge segments
   which are too small.  Finally, all segment breaks too close to the
   start line are removed
}
\seealso{
\code{\link{read_ride}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
