% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_read.R
\name{read_ride}
\alias{read_ride}
\title{read and clean a gps data file}
\usage{
read_ride(
  ridefile,
  tz,
  stopSpeed = 0,
  fixDistance = FALSE,
  loud = FALSE,
  loudSegment = FALSE,
  usefitdc = FALSE,
  lutzmethod = "fast",
  ...
)
}
\arguments{
\item{ridefile}{a filenames (.fit or .gpx extension) to process}

\item{tz}{string containing the timezone for the track data frame}

\item{stopSpeed}{- speed in m/s below which the bike is considered stopped
for time and cadence calculations}

\item{fixDistance}{repair nonmonotonicities in distance which are on
segment breaks - this occurs when power is lost or on some device lockups}

\item{loud}{print information about hr/cadence data issues/fixes}

\item{loudSegment}{print information about re/segmenting track data}

\item{usefitdc}{use package fitdc to read fit files instead of fitparse}

\item{lutzmethod}{method to use to locate timezone, see package lutz}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a list of three data tibbles:  \eqn{summary} and \eqn{trackpoints}
   and \eqn{session}
   These are linked by integer fields \eqn{startbutton.date}
   and \eqn{startbutton.time}
}
\description{
\code{read_ride}  processes a gps .fit and/or .gpx track file to create
  summary and detail data tibbles for modeling or graphical summary.  Track
  segments are adjusted to remove many false autostop-start and startline
  delay sequences.  Several measures of cadence are calculated (including
  zeros, excluding zeros, and midsegment),spurious and missing cadence values
  are repaired, and summary stats are calculated for analysis.
}
\seealso{
\code{\link{read_ridefiles}},
   \code{\link{processSegments}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
