% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_read.R
\name{read_ridefiles}
\alias{read_ridefiles}
\title{read multiple gps data files}
\usage{
read_ridefiles(ridefilevec, cores = 4, loud = FALSE, ...)
}
\arguments{
\item{ridefilevec}{a vector of filenames to process}

\item{cores}{number of cores (default is #CPUs - 1)}

\item{loud}{display summary of re/segmenting actions}

\item{...}{parameters passed for track cleaning}
}
\value{
a list of two data tibbles:  \eqn{summaries} and \eqn{tracks}
   These are linked by integer fields \eqn{startbutton.date}
   and \eqn{startbutton.time}
}
\description{
\code{read_ridefiles} reads a vector of gps .fit and/or .gpx track files
 and put into summary and detail data tibbles for modeling or graphical
 summary.  A wrapper for \code{\link{read_ride}}.
}
\seealso{
\code{\link{read_ride}}
}
