% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadence.R
\name{repairCadence}
\alias{repairCadence}
\title{clean up cadence data for a track}
\usage{
repairCadence(
  trackdf,
  cadNAtoZeroInit = "stopped",
  cadNAtoZeroMid = 3,
  cadNAtoZeroFinal = "downhill",
  fixCadence = TRUE,
  cadMax = 160,
  cadMin = 0,
  cadCorrectTooHigh = "smooth",
  cadCorrectNA = "zero",
  cadCorrectWindowSec = 7,
  cadStuckMax = 0,
  cadStuckRep = 4,
  cadStuckSpdDelta = 0.07,
  cadCorrectStopped = TRUE,
  loud = FALSE,
  ...
)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{cadNAtoZeroInit}{either "all", "stopped","downhill" or "none" to
specify when to replace inital missing cadence values with zero}

\item{cadNAtoZeroMid}{an integer, set midride missing cadence values to zero
if preceded by this many zeros}

\item{cadNAtoZeroFinal}{either "all", "stopped","downhill" or "none" to
specify when to replace final missing cadence values with zero}

\item{fixCadence}{repair cadence errors}

\item{cadMax}{max credible cadence value, larger values are errors}

\item{cadMin}{min nonzero cadence value, smaller values set to 0}

\item{cadCorrectTooHigh}{"smooth" to repair excessive cadence values
using triangular-kernel-weighted average of the nearest nonmissing values
in the same segment or "cap" to cap them at cadMax}

\item{cadCorrectNA}{"smooth" to repair cadence missing values using
triangular-kernel-weighted average of the nearest nonmissing values in
the same segment or "zero" to set them to zero}

\item{cadCorrectWindowSec}{window size for kernel smoothing of cadence}

\item{cadStuckMax}{threshold cadence value for removing repeated low values,
useful for some sensor/GPS combinations.  0 means no checking for this.}

\item{cadStuckRep}{minimum length of runs of low cadence values to remove,
useful for some sensor/GPS combinations}

\item{cadStuckSpdDelta}{proportionate change from initial speed to remove
useful for some sensor/GPS combinations}

\item{cadCorrectStopped}{repair cadence by setting cadence to 0 when speed
is zero.  This is what would be appropriate if magnet parked near sensor
was generating spurious clicks or if moving pedal while at stoplight}

\item{loud}{display actions taken}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairSensorDropOut}},
\code{\link{repairHR}},
\code{\link{repairPower}},
\code{\link{statsHeartRate}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}},
\code{\link{statsTemp}}}
}
\value{
dataframe with cadence data repaired
}
\description{
\code{repairCadence}  processes a gps track file to correct and
  summarize cadence data
}
\seealso{
\code{\link{read_ride}},
   \code{\link{repairSensorDropOut}},
   \code{\link{repairHR}},
   \code{\link{repairPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}},
   \code{\link{statsTemp}}
}
