% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{repairHR}
\alias{repairHR}
\title{clean up HR data for a track}
\usage{
repairHR(trackdf, fixHR = TRUE, HRMax = 220, loud = FALSE, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{fixHR}{repair excessive HR values by setting them to NA}

\item{HRMax}{max credible HR value, larger values are errors set to NA}

\item{loud}{display actions taken}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
dataframe with HR data repaired
}
\description{
\code{repairHR}  processes a gps track file to correct HR data
}
\seealso{
\code{\link{read_ride}},
   \code{\link{statsHeartRate}}
}
