% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{repairPower}
\alias{repairPower}
\title{clean up power data for a track}
\usage{
repairPower(trackdf, powerNAtoZero = TRUE, loud = FALSE, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{powerNAtoZero}{set NA to 0 if true}

\item{loud}{display actions taken}

\item{...}{parameters for \code{\link{excludeCalibrate}},
\code{\link{repairSensorDropOut}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{statsHeartRate}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}},
\code{\link{statsTemp}}}
}
\value{
dataframe with power data repaired
}
\description{
\code{repairPower}  processes a gps track file to correct and
  summarize power data
}
\seealso{
\code{\link{read_ride}},
   \code{\link{repairSensorDropOut}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{repairPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}},
   \code{\link{statsTemp}}
}
