% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{repairSensorDropOut}
\alias{repairSensorDropOut}
\title{clean up sensor dropout data for a track}
\usage{
repairSensorDropOut(trackdf, fixAllSensorDrop = TRUE,
  fixSpeedSensorDrop = TRUE, sensorDropSmooth = TRUE,
  lookBackward = 3, lookForward = 5, loud = FALSE, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{fixAllSensorDrop}{logical, if TRUE fix case of all sensor dropouts}

\item{fixSpeedSensorDrop}{logical, if TRUE fix case of speed sensor dropouts}

\item{sensorDropSmooth}{fill in values}

\item{lookBackward}{number of seconds to look before dropout when fixing}

\item{lookForward}{number of seconds to look after dropout when fixing}

\item{loud}{display actions taken}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
dataframe with data repaired
}
\description{
\code{repairSensorDropOut}  processes a gps track file for data dropouts
}
\seealso{
\code{\link{read_ride}}
}
