% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rideReadGPS.R
\name{rideReadGPS}
\alias{rideReadGPS}
\title{rideReadGPS: A package for reading cycling .fit and .gpx files into
   dataframes containing (corrected) trackpoints and summary statistics}
\description{
rideReadGPS: A package for reading cycling .fit and .gpx files into
   dataframes containing (corrected) trackpoints and summary statistics
}
\section{import gps data}{

  \link{read_ridefiles}, \link{read_ride}
}

