% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracktimes.R
\name{rollingTime}
\alias{rollingTime}
\title{return moving time for a track}
\usage{
rollingTime(trackdf, stopSpeed = 0, include, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{stopSpeed}{a numeric value below which the bike is considered stopped}

\item{include}{a vector specifying which points on the track to include}

\item{...}{parameters for other functions}
}
\value{
a vector of length one containing the elapsed time on the track for which
   include is TRUE
}
\description{
\code{rollingTime}  return the total time in a track
}
\seealso{
\code{\link{read_ride}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{totalTime}},
   \code{\link{pedalTime}}
}
