% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadence.R
\name{statsCadence}
\alias{statsCadence}
\title{summarize cadence statistics for a track}
\usage{
statsCadence(
  trackdf,
  sessionpedalstrokes = NA,
  loud = FALSE,
  cadTrimBegSecs = 15,
  cadTrimBegMeters = 10,
  cadTrimEndSecs = 20,
  cadTrimEndMeters = 15,
  ...
)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{sessionpedalstrokes}{number of pedalstrokes in session records of
fit file}

\item{loud}{print useful trace information}

\item{cadTrimBegSecs}{number of seconds after the beginning of a
track segment that are ignored in calculating midsegment avg cadence}

\item{cadTrimBegMeters}{number of meters after the beginning of a
track segment that are ignored in calculating midsegment avg cadence}

\item{cadTrimEndSecs}{number of seconds before the end of a
track segment that are ignored in calculating midsegment avg cadence}

\item{cadTrimEndMeters}{number of meters before the end of a
track segment that are ignored in calculating midsegment avg cadence}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairSensorDropOut}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{repairPower}},
\code{\link{statsHeartRate}},
\code{\link{statsPower}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}},
\code{\link{statsTemp}}}
}
\value{
a list containing summary data
}
\description{
\code{statsCadence}  processes a gps track file to correct and
  summarize cadence data
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{repairSensorDropOut}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{repairPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsPower}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}},
   \code{\link{statsTemp}}
}
