% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearratio.R
\name{statsGearing}
\alias{statsGearing}
\title{generate gear ratio statistics for a track}
\usage{
statsGearing(trackdf, grLow1 = 0.044, grLow2, minGRCadence = 30, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{grLow1}{lowest gear speed/cadence upper bound}

\item{grLow2}{alternate lowest gear speed/cadence upper bound}

\item{minGRCadence}{ignore where cadence is below this number}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairSensorDropOut}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{repairPower}},
\code{\link{statsHeartRate}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}},
\code{\link{statsTemp}}}
}
\value{
a list containing summary data
}
\description{
\code{statsGearing}  processes a gps track file to summarize the gear ratio
  implied by the ratio of cadence to speed
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{repairSensorDropOut}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{repairPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}},
   \code{\link{statsTemp}}
}
