% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{statsHeartRate}
\alias{statsHeartRate}
\title{generate heartrate statistics for a track}
\usage{
statsHeartRate(trackdf, recovery_hr, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{recovery_hr}{data frame or tibble with gps recovery hr data}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairSensorDropOut}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{repairPower}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}},
\code{\link{statsTemp}}}
}
\value{
a list containing summary data
}
\description{
\code{statsHeartRate}  processes a gps track file to summarize the hr data
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{repairSensorDropOut}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{repairPower}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}},
   \code{\link{statsTemp}}
}
