% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{statsPower}
\alias{statsPower}
\title{generate power statistics for a track}
\usage{
statsPower(trackdf, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data
powerCalibrateTime <= 0}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairSensorDropOut}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{repairPower}},
\code{\link{statsHeartRate}},
\code{\link{statsCadence}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}},
\code{\link{statsTemp}}}
}
\value{
a list containing summary data
}
\description{
\code{statsPower}  processes a gps track file to summarize the power data
}
\seealso{
\code{\link{read_ride}},\code{\link{excludeCalibrate}},
   \code{\link{repairSensorDropOut}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{repairPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsCadence}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}},
   \code{\link{statsTemp}}
}
