% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{statsStops}
\alias{statsStops}
\title{generate stops statistics for a track}
\usage{
statsStops(trackdf, segInitIdleAggSecs = 60, segInitIdleAggMeters = 20,
  ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{segInitIdleAggSecs}{stops/restarts before this time has
elapsed are discarded and the ride starts at the last start prior to this}

\item{segInitIdleAggMeters}{stops/restarts within this distance
of the initial start are discarded as above}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}}}
}
\value{
a list containing summary data
}
\description{
\code{statsStops}  processes a gps track file to summarize start/stop data
}
\seealso{
\code{\link{read_ride}},
   \code{\link{processSegments}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}}
}
