% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempC.R
\name{statsTemp}
\alias{statsTemp}
\title{generate power statistics for a track}
\usage{
statsTemp(trackdf, tempTimeIgnore = 600, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{tempTimeIgnore}{number of seconds after start to ignore temp data}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{statsCadence}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a list containing summary data
}
\description{
\code{statsTemp}  processes a gps track file to summarize the temperature data
}
\seealso{
\code{\link{read_ride}},\code{\link{excludeCalibrate}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{statsCadence}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
