% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracktimes.R
\name{totalTime}
\alias{totalTime}
\title{return the total time for a track}
\usage{
totalTime(trackdf, include, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{include}{a vector specifying which points on the track tyop include}

\item{...}{parameters for other functions}
}
\value{
a vector of length one containing the elapsed time on the track for which
   include is TRUE
}
\description{
\code{totalTime}  return the total time in a track
}
\seealso{
\code{\link{read_ride}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{rollingTime}},
   \code{\link{pedalTime}}
}
