% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{tracks data frame contents}
\alias{tracks data frame contents}
\title{trackpoint data for track files read by bikeCadHr functions}
\description{
trackpoint data for track files read by bikeCadHr functions
}
\details{
Data fields for each trackpoint:

\itemize{
 \item{timestamp.s - POSIXct timestamp}
 \item{distance.m - distance from speed sensor or gps coordinates}
 \item{altitude.m - elevation at trackpoint as reported by gps}
 \item{speed.m.s - speed at trackpoint either speed sensor or gps coordinates}
 \item{heart_rate.bpm - heartrate recorded at the trackpoint}
 \item{cadence.rpm - (possibly corrected) cadence recorded at the trackpoint}
 \item{power.watts - power at trackpoint as reported to GPS from power meter}
 \item{temperature - temperature recorded at the trackpoint}
 \item{segment - segment identifier for the trackpoint}
 \item{stopped - not moving at trackpoint}
 \item{deltatime - numeric time since last trackpoint (seconds)}
 \item{position_lat.dd - latitude in decimal degrees at the trackpoint}
 \item{position_lon.dd - longitude in decimal degrees at the trackpoint}
 \item{startbutton.date - integer containing human readable date stamp for ride,
      used to match summary and detailed track data, derived from the source
      filename (garmin defaults name activity records based on start button time)
      format is YYYYMMDD}
 \item{startbutton.time  - integer containing human readable time stamp for ride,
      used to match summary and detailed track data, derived from the source
      filename (garmin defaults name activity records based on start button time)
      format is  HHMMSS, where HH is 00-23, MM and SS are 00-59}
 }

 \strong{(all distances are in meters, times are in seconds or minutes}
}
