% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{repairPower}
\alias{repairPower}
\title{clean up power data for a track}
\usage{
repairPower(trackdf, loud = FALSE, ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{loud}{display actions taken}

\item{...}{parameters for \code{\link{excludeCalibrate}},
\code{\link{repairHR}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsGrade}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
dataframe with power data repaired
}
\description{
\code{repairPower}  processes a gps track file to correct and
  summarize power data
}
\seealso{
\code{\link{read_ride}},
   \code{\link{statsCadence}}
}
