% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{statsGrade}
\alias{statsGrade}
\title{generate grade statistics for a track}
\usage{
statsGrade(trackdf, gradeSmoothWindowMeters = 7,
  gradeSmoothWindowNN = 10, flatLimit = 0.01, gradeMinDistance = 0.5,
  ...)
}
\arguments{
\item{trackdf}{data frame or tibble with gps track data}

\item{gradeSmoothWindowMeters}{bandwidth (meters) for kernel smoothing of
elevation in the grade calculation}

\item{gradeSmoothWindowNN}{number of observations on each side used for
calculating kernel smoothed elevation for the grade calculation}

\item{flatLimit}{if the grade does not exceed this, it is neither scending
nor descending}

\item{gradeMinDistance}{if the distance (meters) between points is less than
this,the distance will not be counted aas ascending or descending and the
grade will not be calculated}

\item{...}{parameters for \code{\link{processSegments}},
\code{\link{repairHR}},
\code{\link{repairCadence}},
\code{\link{statsCadence}},
\code{\link{statsPower}},
\code{\link{statsHeartRate}},
\code{\link{statsGearing}},
\code{\link{statsSession}},
\code{\link{statsStops}}}
}
\value{
a list containing summary data
}
\description{
\code{statsGrade}  processes a gps track file to summarize grade data
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsSession}},
   \code{\link{statsStops}}
}
