% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{statsSession}
\alias{statsSession}
\title{collect gps session statistics for a track}
\usage{
statsSession(session)
}
\arguments{
\item{session}{data frame or tibble with gps session data}
}
\value{
a list containing summary data
}
\description{
\code{statsSession}  processes a gps session file
}
\seealso{
\code{\link{read_ride}},\code{\link{processSegments}},
   \code{\link{repairHR}},
   \code{\link{repairCadence}},
   \code{\link{statsCadence}},
   \code{\link{statsPower}},
   \code{\link{statsHeartRate}},
   \code{\link{statsGearing}},
   \code{\link{statsGrade}},
   \code{\link{statsStops}}
}
